fix/add image respone and fix delete comment

+++ src/main/java/com/vietqr/org/service/qrfeed/QrCommentServiceImpl.java
@@ -3,7 +3,10 @@ package com.vietqr.org.service.qrfeed;
 import com.vietqr.org.dto.qrfeed.QrCommentDTO;
 import com.vietqr.org.dto.qrfeed.QrCommentRequestDTO;
 import com.vietqr.org.dto.qrfeed.UserCommentDTO;
+import com.vietqr.org.entity.qrfeed.QrCommentEntity;
+import com.vietqr.org.entity.qrfeed.QrWalletEntity;
 import com.vietqr.org.repository.QrCommentRepository;
+import com.vietqr.org.repository.QrWalletRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +17,9 @@ import java.util.UUID;
 public class QrCommentServiceImpl implements QrCommentService{
     @Autowired
     private QrCommentRepository qrCommentRepository;
+
+    @Autowired
+    private QrWalletRepository qrWalletRepository;
 
 
     @Override
@@ -31,9 +37,22 @@ public class QrCommentServiceImpl implements QrCommentService{
     }
 
     @Override
-    public void deleteComment(String commentId) {
-        qrCommentRepository.unlinkCommentFromQrWallet(commentId);
-        qrCommentRepository.deleteComment(commentId);
+    public boolean deleteComment(String commentId, String userId) {
+        QrCommentEntity comment = qrCommentRepository.findById(commentId).orElse(null);
+        if (comment == null) {
+            return false;
+        }
+        String qrWalletId = qrCommentRepository.findQrWalletIdByCommentId(commentId);
+        QrWalletEntity qrWallet = qrWalletRepository.findById(qrWalletId).orElse(null);
+        if (qrWallet == null) {
+            return false;
+        }
+        if (comment.getUserId().equals(userId) || qrWallet.getUserId().equals(userId)) {
+            qrCommentRepository.unlinkCommentFromQrWallet(commentId);
+            qrCommentRepository.deleteComment(commentId);
+            return true;
+        }
+        return false;
     }
 
     @Override
